requirements.txt
pandas
tabulate
ibis-framework[duckdb,polars,sqlite]
ipython
